# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Hypermedia project by Roccamena, Romanò, Buttolo"
  version: "1.0.0"
  title: "Child4Help"
  termsOfService: "http://swagger.io/terms/"
host: "localhost:8080"
basePath: "/api"
...
schemes:
- "https"
- "http"
paths:
...
  /message/{id_message}:
    get:
      tags:
      - "message"
      summary: "Get a message by ID"
      operationId: "getMessageByID"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_message"
        in: "path"
        description: "ID of message to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Invalid ID"
        "404":
          description: "Message not found"
      x-swagger-router-controller: "Message"
    delete:
      tags:
      - "message"
      summary: "Delete a message by ID"
      operationId: "deleteMessage"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_message"
        in: "path"
        description: "ID of message to be deleted"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid ID"
        "404":
          description: "Message not found"
      x-swagger-router-controller: "Message"
  /message:
    post:
      tags:
      - "message"
      summary: "Add a new message"
      operationId: "addMessage"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "message"
        description: "Message object to be added"
        required: true
        schema:
          $ref: "#/definitions/Message"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid input"
      x-swagger-router-controller: "Message"
  /link/related/service:
    post:
      tags:
      - "link"
      summary: "Add new related relation beetween services"
      operationId: "addRelatedServices"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_service1"
        in: "query"
        description: "ID of first service"
        required: true
        type: "integer"
        format: "int64"
      - name: "id_service2"
        in: "query"
        description: "ID of second service"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid operation"
        "404":
          description: "Services not found"
      x-swagger-router-controller: "Link"
    delete:
      tags:
      - "link"
      summary: "Delete related relation beetween services"
      operationId: "deleteRelatedServices"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_service1"
        in: "query"
        description: "ID of first service"
        required: true
        type: "integer"
        format: "int64"
      - name: "id_service2"
        in: "query"
        description: "ID of second service"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid operation"
        "404":
          description: "Services not found"
      x-swagger-router-controller: "Link"
  /link/related/event:
    post:
      tags:
      - "link"
      summary: "Add new related relation beetween events"
      operationId: "addRelatedEvents"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_event1"
        in: "query"
        description: "ID of first event"
        required: true
        type: "integer"
        format: "int64"
      - name: "id_event2"
        in: "query"
        description: "ID of second event"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid operation"
        "404":
          description: "Events not found"
      x-swagger-router-controller: "Link"
    delete:
      tags:
      - "link"
      summary: "Delete related relation beetween events"
      operationId: "deleteRelatedEvents"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_event1"
        in: "query"
        description: "ID of first event"
        required: true
        type: "integer"
        format: "int64"
      - name: "id_event2"
        in: "query"
        description: "ID of second event"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid operation"
        "404":
          description: "Events not found"
      x-swagger-router-controller: "Link"
  /link/related/article:
    post:
      tags:
      - "link"
      summary: "Add new related relation beetween articles"
      operationId: "addRelatedArticles"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_article1"
        in: "query"
        description: "ID of first article"
        required: true
        type: "integer"
        format: "int64"
      - name: "id_article2"
        in: "query"
        description: "ID of second article"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid operation"
        "404":
          description: "Articles not found"
      x-swagger-router-controller: "Link"
    delete:
      tags:
      - "link"
      summary: "Delete related relation beetween articles"
      operationId: "deleteRelatedArticles"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_article1"
        in: "query"
        description: "ID of first article"
        required: true
        type: "integer"
        format: "int64"
      - name: "id_article2"
        in: "query"
        description: "ID of second article"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid operation"
        "404":
          description: "Articles not found"
      x-swagger-router-controller: "Link"
  /link/presented:
    post:
      tags:
      - "link"
      summary: "Add new service presented in event relation"
      operationId: "addServicesEvents"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_service"
        in: "query"
        description: "ID of service"
        required: true
        type: "integer"
        format: "int64"
      - name: "id_event"
        in: "query"
        description: "ID of event"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid operation"
        "404":
          description: "Service or event not found"
      x-swagger-router-controller: "Link"
    delete:
      tags:
      - "link"
      summary: "Delete service presented in event relation"
      operationId: "deleteServicesEvents"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_service"
        in: "query"
        description: "ID of service"
        required: true
        type: "integer"
        format: "int64"
      - name: "id_event"
        in: "query"
        description: "ID of event"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid operation"
        "404":
          description: "Service or event not found"
      x-swagger-router-controller: "Link"
  /link/involved:
    post:
      tags:
      - "link"
      summary: "Add new person involved in service relation"
      operationId: "addPeopleServices"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_person"
        in: "query"
        description: "ID of person"
        required: true
        type: "integer"
        format: "int64"
      - name: "id_service"
        in: "query"
        description: "ID of service"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid operation"
        "404":
          description: "Person or service not found"
      x-swagger-router-controller: "Link"
    delete:
      tags:
      - "link"
      summary: "Delete person involved in service relation"
      operationId: "deletePeopleServices"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_person"
        in: "query"
        description: "ID of person"
        required: true
        type: "integer"
        format: "int64"
      - name: "id_service"
        in: "query"
        description: "ID of service"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid operation"
        "404":
          description: "Person or service not found"
      x-swagger-router-controller: "Link"
  /link/contact:
    put:
      tags:
      - "link"
      summary: "Update event's contact relation"
      operationId: "updateEventContact"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_event"
        in: "query"
        description: "ID of event"
        required: true
        type: "integer"
        format: "int64"
      - name: "id_person"
        in: "query"
        description: "ID of person"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid operation"
        "404":
          description: "Event or person not found"
      x-swagger-router-controller: "Link"
  /link/discussed/service:
    put:
      tags:
      - "link"
      summary: "Update service discussed in article relation"
      operationId: "updateServiceArticles"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_article"
        in: "query"
        description: "ID of article"
        required: true
        type: "integer"
        format: "int64"
      - name: "id_service"
        in: "query"
        description: "ID of service"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid operation"
        "404":
          description: "Article or service not found"
      x-swagger-router-controller: "Link"
  /link/discussed/event:
    put:
      tags:
      - "link"
      summary: "Update event discussed in article relation"
      operationId: "updateEventArticles"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_article"
        in: "query"
        description: "ID of article"
        required: true
        type: "integer"
        format: "int64"
      - name: "id_event"
        in: "query"
        description: "ID of event"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid operation"
        "404":
          description: "Article or event not found"
      x-swagger-router-controller: "Link"
definitions:
...
  Message:
    type: "object"
    required:
    - "id_message"
    properties:
      id_message:
        type: "integer"
        format: "int64"
      sender_name:
        type: "string"
      sender_email:
        type: "string"
        format: "email"
      title:
        type: "string"
      body:
        type: "string"
    xml:
      name: "Message"
    example:
      id_message: 0
      sender_name: "sender_name"
      title: "title"
      body: "body"
      sender_email: "sender_email"